+++ src/qemu/qemu_monitor.h
@@ -519,9 +519,9 @@ struct _qemuMonitorMigrationStats {
     unsigned long long xbzrle_cache_miss;
     unsigned long long xbzrle_overflow;
 
-    unsigned long long chkpt_size;
-    unsigned long long chkpt_length;
-    unsigned long long chkpt_pause;
+    double chkpt_size;
+    double chkpt_length;
+    double chkpt_pause;
     unsigned long long chkpt_count;
     unsigned long long chkpt_proxy_discompare;
 };
